Improve Robustness and fix error when synchronizing for the first time
Fix backend connection drops, errors when synchronizing for the first time or when transferring the generated payload:
* Previously, `find_one` would fail on new accounts (because there are no entries in the database), and chaining `.unwrap()` panicked the server worker thread. So replaced **every** `unwrap()` iteration with structured pattern matching. For instance, in `sync/settings/service.rs`, if a new user attempts a download-only sync, the backend now safely returns `{"settings": null}` instead of throwing an `InternalServerError()`.
* Add Null safety fallbacks by safely promoted all non-critical client fields to `Option<T>` within Rust, securing the pipeline payload against missing properties.
* Force all 4 sync controllers and 2 user controllers to return `401 Unauthorized` safely.
* Fix i32 overflow -> Use of i64 to avoid error on Unix millisecond timestamps.
* Graceful Degradation applied in Upsert Ops, the server now `continue`s, dropping specifically malformed nodes but saving the rest of the valid user library payload, ensuring the end-user's data retains maximum integrity in `manga/service.rs`, `history/service.rs`, and `update/service.rs`.
* Sanitize deletion -> The backend handles resets securely now by managing `empty_id` references correctly and skipping arbitrary runtime cloning of options that used heavy runtime footprints.

Add Tests:
* Wrote 12 strict JSON data simulations proving boundaries before updating the Rust Model fields logic in `tests/sync_deserialization_tests.rs`.
* Add HTTP Real Simulation File : `tests/api/sync_endpoints.http`

Remove OpenSSL to directly use Rustls which is safe, robust and don't need to have Perl and OpenSSL installed.

+++ src/sync/history/model.rs
@@ -3,14 +3,14 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct History {
-    #[serde(rename="_id", skip_serializing)]
+    #[serde(rename = "_id", skip_serializing)]
     pub oid: Option<ObjectId>,
-    pub id: i32,
+    pub id: i64,
     pub date: String,
     #[serde(rename = "mangaId")]
-    pub manga_id: i32,
+    pub manga_id: i64,
     #[serde(rename = "chapterId")]
-    pub chapter_id: i32,
+    pub chapter_id: i64,
     #[serde(rename = "itemType")]
     pub item_type: i32,
     #[serde(skip_serializing)]
@@ -22,7 +22,7 @@ pub struct History {
 #[derive(Serialize, Deserialize)]
 pub struct HistoryList {
     pub histories: Vec<History>,
-    pub deleted_histories: Vec<i32>,
+    pub deleted_histories: Vec<i64>,
     #[serde(rename = "resetAll")]
     pub reset_all: Option<bool>,
 }
