+++ src/sync/update/model.rs
@@ -5,9 +5,9 @@ use serde::{Deserialize, Serialize};
 pub struct Update {
     #[serde(rename = "_id", skip_serializing)]
     pub oid: Option<ObjectId>,
-    pub id: i32,
+    pub id: i64,
     #[serde(rename = "mangaId")]
-    pub manga_id: i32,
+    pub manga_id: i64,
     #[serde(rename = "chapterName")]
     pub chapter_name: String,
     pub date: String,
@@ -20,7 +20,7 @@ pub struct Update {
 #[derive(Serialize, Deserialize)]
 pub struct UpdateList {
     pub updates: Vec<Update>,
-    pub deleted_updates: Vec<i32>,
+    pub deleted_updates: Vec<i64>,
     #[serde(rename = "resetAll")]
     pub reset_all: Option<bool>,
 }
